2. bugs were fixed.
1. A non-redundant antibody is PDB file is dealt as antigen to check contacts
with the CDRS of other antibody to check if it is idiotypic antibody. If any such antibody has L and H labels, they will be ignored bcoz its is most likely that they won't be idiotypic rather chaincontacts is calculating contacts of LH (antibody) with L and H antigen. But it will return contacts with itself. Therefore any such case will be ignored

2. <15 CDR contacts with antigen won't be considered.